0.7.1 Release and example code
This is the 0.7.1 release that has been pushed to the market which fixes
a bug in the weather refreshing and makes the weather screen display
relevant loading information.
This also includes the "fragment6" unused example fragment and a small
bit of code in fragment4 and its list adapter that let it run code when
the user reaches the end of the list.